google-cloud/generate_ai/cloud_functions/langchain_bigquery/langchain_bigquery_chat_agent/requirements.txt
google-cloud-aiplatform
langchain 
openai 
chromadb 
tiktoken 
tabulate 
sqlalchemy 
sqlalchemy-bigquery 
google-cloud-bigquery
google-cloud-secret-manager
flask
